Update README.md
Paris–Brest–Paris is a long-distance cycling event. It was originally a 1,200 km bicycle race from Paris to Brest and back to Paris in 1891. The last time it was run as a race was 1951. The most recent edition of PBP was held on 20 August 2023. In 1931 amateur cyclists were separated from professionals. Wikipedia